requirements.txt
python-dotenv
django
django-ninja
PyJWT
gunicorn